+++ lib/app.dart
@@ -7,6 +7,7 @@ import 'package:delivery_app/features/auth/shared/presentation/bloc/auth_bloc.da
 import 'package:delivery_app/features/notifications/notification_list/presentation/bloc/notification_bloc.dart';
 import 'package:delivery_app/features/settings/presentation/cubit/settings_cubit.dart';
 import 'package:delivery_app/features/settings/presentation/cubit/settings_state.dart';
+import 'package:delivery_app/features/trips/trip_list/presentation/bloc/trip_list_bloc.dart';
 import 'package:delivery_app/injection_container.dart';
 import 'package:delivery_app/shared/widgets/banners/offline_banner.dart';
 import 'package:easy_localization/easy_localization.dart';
@@ -27,6 +28,7 @@ class _NoktaAppState extends State<NoktaApp> {
     super.initState();
     WidgetsBinding.instance.addPostFrameCallback((_) {
       sl<NotificationBloc>().add(const NotificationLoadRequested());
+      sl<TripListBloc>().add(const TripListLoadRequested());
     });
   }
 
@@ -38,6 +40,7 @@ class _NoktaAppState extends State<NoktaApp> {
         BlocProvider<AuthBloc>.value(value: sl<AuthBloc>()),
         BlocProvider<SettingsCubit>.value(value: sl<SettingsCubit>()),
         BlocProvider<NotificationBloc>.value(value: sl<NotificationBloc>()),
+        BlocProvider<TripListBloc>.value(value: sl<TripListBloc>()),
       ],
       child: BlocListener<SettingsCubit, SettingsState>(
         listenWhen: (prev, curr) => prev.locale != curr.locale,